Refactor peripheral configuration loading
